% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FillMissingValues.R
\name{FillMissingValues}
\alias{FillMissingValues}
\title{Fill missing values}
\usage{
FillMissingValues(df, maxgap)
}
\arguments{
\item{df}{dataframe containing the timeseries in columns separated by ID (Header must follow this convention: column 1 = "datetime", column 2 = "SiteID", column 3 = "variable name"). df can be the result of GetDataFromECMWF().}

\item{maxgap}{maximum gap to interpolate (e.g. 6 hours)}
}
\value{
updated df with infilled values
}
\description{
Fill missing values
}
\examples{
# FillMissingValues(clima, maxgap = 6)

}

