% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataFromECMWF.R
\name{GetDataFromECMWF}
\alias{GetDataFromECMWF}
\title{Get data from ECMWF ERA_Interim}
\usage{
GetDataFromECMWF(points, years = 1981:2014, var = "t2m", timestep = 6,
  prefix = "", path = "~")
}
\arguments{
\item{points}{are lat/lon coordinates of points (e.g. stations)}

\item{years}{years to retrieve data for}

\item{var}{variable to retrieve}

\item{timestep}{time step in hours (e.g. 1, 3, 6, 12, 24)}

\item{prefix}{string starting netcdf file name}

\item{path}{folder path where netcdf files are stored}
}
\value{
time series variable
}
\description{
Get data from ECMWF ERA_Interim
}
\details{
Possible variables names are: "t2m" (2m temperature, in K), "u10" (10 metres wind U component, in m/s), "v10" (10 metres wind V component, in m/s), "tp" (total precipitation, in m), "blh" (boundary layer height, in m), "ssr" (surface net solar radiation, in W/m2s).
}
\examples{
# GetDataFromECMWF(points, years = 1981:2014, var = "t2m",
#                  timestep = 6, prefix = "", path = "~")

}

