% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSeason.R
\name{getSeason}
\alias{getSeason}
\title{Get season a date belongs to}
\usage{
getSeason(DATES)
}
\arguments{
\item{DATES}{a date.}
}
\value{
returns the name of the season (e.g. "Fall")
}
\description{
Get season a date belongs to. This function was taken from the following stackoverflow post: http://stackoverflow.com/questions/9500114/find-which-season-a-particular-date-belongs-to.
}
\examples{
# my.dates <- as.Date("2011-12-01", format = "\%Y-\%m-\%d") + 0:60
# getSeason(my.dates)

}

