\name{kehra-package}
\alias{kehra-package}
\alias{kehra}
\docType{package}
\title{
\packageTitle{kehra}
}
\description{
\packageDescription{kehra}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{kehra}
\packageIndices{kehra}
Collection of utility functions used in the KEHRA project
}
\author{
\packageAuthor{kehra}
Maintainer: \packageMaintainer{kehra}
}

\keyword{ package }
