% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windDir.R
\name{windDir}
\alias{windDir}
\title{Wind Direction}
\usage{
windDir(u, v)
}
\arguments{
\item{u}{first component of wind speed}

\item{v}{second component of wind speed}
}
\value{
direction in degrees from u & v components
}
\description{
Calculate wind direction in degrees from u & v components
}
\examples{
# windDir(u, v)

}

