% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windDirection.R
\name{windDirection}
\alias{windDirection}
\title{Wind Direction}
\usage{
windDirection(u, v)
}
\arguments{
\item{u}{first component of wind speed}

\item{v}{second component of wind speed}
}
\value{
direction in degrees from u & v components
}
\description{
Calculate wind direction in degrees from u & v components
}
\examples{
# windDirection(u, v)

}

