% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windSpeed.R
\name{windSpeed}
\alias{windSpeed}
\title{Wind Speed}
\usage{
windSpeed(u, v)
}
\arguments{
\item{u}{first component of wind speed}

\item{v}{second component of wind speed}
}
\value{
Speed in m/s
}
\description{
Calculate wind speed in m/s from u & v components
}
\examples{
# windSpeed(u, v)

}

