% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillMissingValues.R
\name{fillMissingValues}
\alias{fillMissingValues}
\title{Fill missing values}
\usage{
fillMissingValues(id, df, maxgap = 12, parallel = FALSE)
}
\arguments{
\item{id}{site id code}

\item{df}{dataframe containing the timeseries in columns separated by ID (Header must follow this convention: column 1 = "datetime", column 2 = "SiteID", column 3 = "variable name"). df can be the result of GetDataFromECMWF().}

\item{maxgap}{maximum gap to interpolate (e.g. 6 hours)}

\item{parallel}{Bolean, if TRUE parallel jobs are allowed.}
}
\value{
updated df with infilled values
}
\description{
Fill missing values
}
\examples{
# fillMissingValues(clima)

}

