% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDataFromECMWF.R
\name{GetDataFromECMWF}
\alias{GetDataFromECMWF}
\title{Get data from ECMWF ERA_Interim}
\usage{
GetDataFromECMWF(db, points, years = 1981:2014, var = "t2m", prefix = "",
  timestep = 6)
}
\arguments{
\item{db}{dataframe to populate with info from ERA-Interim}

\item{points}{are lat/lon coordinates of points (e.g. stations)}

\item{years}{years to retrieve data for}

\item{var}{variable to retrieve}

\item{prefix}{string starting netcdf file name}

\item{timestep}{time step in hours (e.g. 1, 3, 6, 12, 24)}
}
\value{
updated db
}
\description{
Get data from ECMWF ERA_Interim
}
\details{
Possible variables names are: "t2m" (2m temperature, in K), "u10" (10 metres wind U component, in m/s), "v10" (10 metres wind V component, in m/s), "tp" (total precipitation, in m), "blh" (boundary layer height, in m), "ssr" (surface net solar radiation, in W/m2s).
}
\examples{
# GetDataFromECMWF(db, points, years = 1981:2014, var = "t2m", prefix = "", timestep = 6)

}

